+++ src/gwt/src/org/rstudio/studio/client/workbench/commands/Commands.java
@@ -593,6 +593,8 @@ public abstract class
    public abstract AppCommand openDeveloperConsole();
    public abstract AppCommand reloadUi();
    public abstract AppCommand showGpuDiagnostics();
+   public abstract AppCommand toggleGpuExclusionList();
+   public abstract AppCommand toggleGpuDriverBugWorkarounds();
    public abstract AppCommand showA11yDiagnostics();
    public abstract AppCommand showLogFiles();
    public abstract AppCommand rstudioSupport();